refactor logic into the controllers directory
