+++ server/controllers/tasks.js
@@ -1,0 +1,93 @@
+const mongoose = require('mongoose'),
+  Task = require('../db/models/task');
+
+exports.createTask = async (req, res) => {
+  const task = await new Task({
+    ...req.body,
+    owner: req.user._id
+  });
+  try {
+    task.save();
+    res.status(201).json(task);
+  } catch (e) {
+    res.status(400).json({ error: e.toString() });
+  }
+};
+
+exports.getSpecificTask = async (req, res) => {
+  const _id = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(_id))
+    return res.status(400).send('Not a valid task id');
+
+  try {
+    const task = await Task.findOne({ _id, owner: req.user._id });
+    if (!task) return res.status(404).send();
+
+    res.json(task);
+  } catch (e) {
+    res.status(500).json({ error: e.toString() });
+  }
+};
+
+exports.getAllTasks = async (req, res) => {
+  const match = {},
+    sort = {};
+
+  if (req.query.completed) match.completed = req.query.completed === 'true';
+
+  if (req.query.sortBy) {
+    const parts = req.query.sortBy.split(':');
+    sort[parts[0]] = parts[1] === 'desc' ? -1 : 1;
+  }
+  try {
+    await req.user
+      .populate({
+        path: 'tasks',
+        match,
+        options: {
+          limit: parseInt(req.query.limit),
+          skip: parseInt(req.query.skip),
+          sort
+        }
+      })
+      .execPopulate();
+    res.json(req.user.tasks);
+  } catch (e) {
+    res.status(500).json({ error: e.toString() });
+  }
+};
+
+exports.updateTask = async (req, res) => {
+  const updates = Object.keys(req.body);
+  const allowedUpdates = ['description', 'completed', 'dueDate'];
+  const isValidOperation = updates.every((update) =>
+    allowedUpdates.includes(update)
+  );
+  if (!isValidOperation)
+    return res.status(400).send({ error: 'Invalid updates!' });
+  try {
+    const task = await Task.findOne({
+      _id: req.params.id,
+      owner: req.user._id
+    });
+    if (!task) return res.status(404).json({ error: 'task not found' });
+    updates.forEach((update) => (task[update] = req.body[update]));
+    await task.save();
+    res.json(task);
+  } catch (e) {
+    res.status(400).json({ error: e.toString() });
+  }
+};
+
+exports.deleteTask = async (req, res) => {
+  try {
+    const task = await Task.findOneAndDelete({
+      _id: req.params.id,
+      owner: req.user._id
+    });
+    if (!task) return res.status(404).json({ error: 'task not found' });
+    res.json(task);
+  } catch (e) {
+    res.status(500).json({ error: e.toString() });
+  }
+};
