+++ server/controllers/users.js
@@ -1,0 +1,163 @@
+const User = require('../db/models/user'),
+  {
+    sendCancellationEmail,
+    sendWelcomeEmail,
+    forgotPasswordEmail
+  } = require('../emails/index'),
+  cloudinary = require('cloudinary').v2,
+  jwt = require('jsonwebtoken');
+
+//Functions for OPEN Routes
+exports.createUser = async (req, res) => {
+  const { name, email, password } = req.body;
+  try {
+    const user = new User({
+      name,
+      email,
+      password
+    });
+
+    const token = await user.generateAuthToken();
+    res.cookie('jwt', token, {
+      httpOnly: true,
+      sameSite: 'Strict',
+      secure: process.env.NODE_ENV !== 'production' ? false : true
+    });
+    sendWelcomeEmail(user.email, user.name);
+    res.status(201).json(user);
+  } catch (e) {
+    res.status(400).json({ error: e.toString() });
+  }
+};
+
+exports.loginUser = async (req, res) => {
+  const { email, password } = req.body;
+  try {
+    const user = await User.findByCredentials(email, password);
+    const token = await user.generateAuthToken();
+    res.cookie('jwt', token, {
+      httpOnly: true,
+      sameSite: 'Strict',
+      secure: process.env.NODE_ENV !== 'production' ? false : true
+    });
+    res.json(user);
+  } catch (e) {
+    res.status(400).json({ error: e.toString() });
+  }
+};
+
+exports.requestPasswordReset = async (req, res) => {
+  try {
+    const { email } = req.query,
+      user = await User.findOne({ email });
+    if (!user) throw new Error("account doesn't exist");
+    // Build jwt token
+    const token = jwt.sign(
+      { _id: user._id.toString(), name: user.name },
+      process.env.JWT_SECRET,
+      {
+        expiresIn: '10m'
+      }
+    );
+    forgotPasswordEmail(email, token);
+    res.json({ message: 'reset password email sent' });
+  } catch (e) {
+    res.json({ error: e.toString() });
+  }
+};
+
+exports.passwordRedirect = async (req, res) => {
+  const { token } = req.params;
+  try {
+    jwt.verify(token, process.env.JWT_SECRET, function (err, decoded) {
+      if (err) throw new Error(err.message);
+    });
+    res.cookie('jwt', token, {
+      httpOnly: true,
+      maxAge: 600000,
+      sameSite: 'Strict'
+    });
+    res.redirect(process.env.URL + '/update-password');
+  } catch (e) {
+    res.json({ error: e.toString() });
+  }
+};
+
+//Functions for SECURE Routes
+exports.getCurrentUser = async (req, res) => res.json(req.user);
+
+exports.updateCurrentUser = async (req, res) => {
+  const updates = Object.keys(req.body);
+  const allowedUpdates = ['name', 'email', 'password', 'avatar'];
+  const isValidOperation = updates.every((update) =>
+    allowedUpdates.includes(update)
+  );
+  if (!isValidOperation)
+    return res.status(400).send({ error: 'invalid updates!' });
+  try {
+    updates.forEach((update) => (req.user[update] = req.body[update]));
+    await req.user.save();
+    res.json(req.user);
+  } catch (e) {
+    res.status(400).json({ error: e.toString() });
+  }
+};
+
+exports.logoutUser = async (req, res) => {
+  try {
+    req.user.tokens = req.user.tokens.filter((token) => {
+      return token.token !== req.token;
+    });
+    await req.user.save();
+    res.clearCookie('jwt');
+    res.json({ message: 'Logged out' });
+  } catch (e) {
+    res.status(500).json({ error: e.toString() });
+  }
+};
+
+exports.logoutAllDevices = async (req, res) => {
+  try {
+    req.user.tokens = [];
+    await req.user.save();
+    res.clearCookie('jwt');
+    res.json({ message: 'all devices logged out' });
+  } catch (e) {
+    res.status(500).send();
+  }
+};
+
+exports.deleteUser = async (req, res) => {
+  try {
+    await req.user.remove();
+    sendCancellationEmail(req.user.email, req.user.name);
+    res.clearCookie('jwt');
+    res.json({ message: 'user deleted' });
+  } catch (e) {
+    res.status(500).json({ error: e.toString() });
+  }
+};
+
+exports.uploadAvatar = async (req, res) => {
+  try {
+    const response = await cloudinary.uploader.upload(
+      req.files.avatar.tempFilePath
+    );
+    req.user.avatar = response.secure_url;
+    await req.user.save();
+    res.json(response);
+  } catch (e) {
+    res.json({ error: e.toString() });
+  }
+};
+
+exports.updatePassword = async (req, res) => {
+  try {
+    req.user.password = req.body.password;
+    await req.user.save();
+    res.clearCookie('jwt');
+    res.json({ message: 'password updated successfully' });
+  } catch (e) {
+    res.json({ error: e.toString() });
+  }
+};
