+++ server/routes/secure/tasks.js
@@ -1,40 +1,21 @@
 const router = require('express').Router(),
-  mongoose = require('mongoose'),
-  Task = require('../../db/models/task');
+  {
+    createTask,
+    getAllTasks,
+    getSpecificTask,
+    updateTask,
+    deleteTask
+  } = require('../../controllers/tasks');
 
 // ***********************************************//
 // Create a task
 // ***********************************************//
-router.post('/', async (req, res) => {
-  const task = await new Task({
-    ...req.body,
-    owner: req.user._id
-  });
-  try {
-    task.save();
-    res.status(201).json(task);
-  } catch (e) {
-    res.status(400).json({ error: e.toString() });
-  }
-});
+router.post('/', createTask);
 
 // ***********************************************//
 // Get a specific task
 // ***********************************************//
-router.get('/:id', async (req, res) => {
-  const _id = req.params.id;
-  if (!mongoose.Types.ObjectId.isValid(_id))
-    return res.status(400).send('Not a valid task id');
-
-  try {
-    const task = await Task.findOne({ _id, owner: req.user._id });
-    if (!task) return res.status(404).send();
-
-    res.json(task);
-  } catch (e) {
-    res.status(500).json({ error: e.toString() });
-  }
-});
+router.get('/:id', getSpecificTask);
 
 // ***********************************************//
 // Get all tasks
@@ -43,72 +24,15 @@ router.get('/:id', async (req, res) => {
 // /tasks?sortBy=createdAt:asc
 // /tasks?sortBy=dueDate:desc
 // ***********************************************//
-router.get('/', async (req, res) => {
-  const match = {},
-    sort = {};
-
-  if (req.query.completed) match.completed = req.query.completed === 'true';
-
-  if (req.query.sortBy) {
-    const parts = req.query.sortBy.split(':');
-    sort[parts[0]] = parts[1] === 'desc' ? -1 : 1;
-  }
-  try {
-    await req.user
-      .populate({
-        path: 'tasks',
-        match,
-        options: {
-          limit: parseInt(req.query.limit),
-          skip: parseInt(req.query.skip),
-          sort
-        }
-      })
-      .execPopulate();
-    res.json(req.user.tasks);
-  } catch (e) {
-    res.status(500).json({ error: e.toString() });
-  }
-});
+router.get('/', getAllTasks);
 // ***********************************************//
 // Update a task
 // ***********************************************//
-router.patch('/:id', async (req, res) => {
-  const updates = Object.keys(req.body);
-  const allowedUpdates = ['description', 'completed', 'dueDate'];
-  const isValidOperation = updates.every((update) =>
-    allowedUpdates.includes(update)
-  );
-  if (!isValidOperation)
-    return res.status(400).send({ error: 'Invalid updates!' });
-  try {
-    const task = await Task.findOne({
-      _id: req.params.id,
-      owner: req.user._id
-    });
-    if (!task) return res.status(404).json({ error: 'task not found' });
-    updates.forEach((update) => (task[update] = req.body[update]));
-    await task.save();
-    res.json(task);
-  } catch (e) {
-    res.status(400).json({ error: e.toString() });
-  }
-});
+router.patch('/:id', updateTask);
 
 // ***********************************************//
 // Delete a task
 // ***********************************************//
-router.delete('/:id', async (req, res) => {
-  try {
-    const task = await Task.findOneAndDelete({
-      _id: req.params.id,
-      owner: req.user._id
-    });
-    if (!task) return res.status(404).json({ error: 'task not found' });
-    res.json(task);
-  } catch (e) {
-    res.status(500).json({ error: e.toString() });
-  }
-});
+router.delete('/:id', deleteTask);
 
 module.exports = router;
