+++ server/routes/secure/users.js
@@ -1,104 +1,47 @@
 const router = require('express').Router(),
-  { sendCancellationEmail } = require('../../emails/index'),
-  cloudinary = require('cloudinary').v2;
+  {
+    getCurrentUser,
+    updateCurrentUser,
+    logoutUser,
+    logoutAllDevices,
+    deleteUser,
+    uploadAvatar,
+    updatePassword
+  } = require('../../controllers/users');
 
 // ***********************************************//
 // Get current user
 // ***********************************************//
-router.get('/me', async (req, res) => res.json(req.user));
+router.get('/me', getCurrentUser);
 
 // ***********************************************//
 // Update a user
 // ***********************************************//
-router.patch('/me', async (req, res) => {
-  const updates = Object.keys(req.body);
-  const allowedUpdates = ['name', 'email', 'password', 'avatar'];
-  const isValidOperation = updates.every((update) =>
-    allowedUpdates.includes(update)
-  );
-  if (!isValidOperation)
-    return res.status(400).send({ error: 'invalid updates!' });
-  try {
-    updates.forEach((update) => (req.user[update] = req.body[update]));
-    await req.user.save();
-    res.json(req.user);
-  } catch (e) {
-    res.status(400).json({ error: e.toString() });
-  }
-});
+router.patch('/me', updateCurrentUser);
 
 // ***********************************************//
 // Logout a user
 // ***********************************************//
-router.post('/logout', async (req, res) => {
-  try {
-    req.user.tokens = req.user.tokens.filter((token) => {
-      return token.token !== req.token;
-    });
-    await req.user.save();
-    res.clearCookie('jwt');
-    res.json({ message: 'Logged out' });
-  } catch (e) {
-    res.status(500).json({ error: e.toString() });
-  }
-});
+router.post('/logout', logoutUser);
 
 // ***********************************************//
 // Logout all devices
 // ***********************************************//
-router.post('/logoutAll', async (req, res) => {
-  try {
-    req.user.tokens = [];
-    await req.user.save();
-    res.clearCookie('jwt');
-    res.json({ message: 'all devices logged out' });
-  } catch (e) {
-    res.status(500).send();
-  }
-});
+router.post('/logoutAll', logoutAllDevices);
 
 // ***********************************************//
 // Delete a user
 // ***********************************************//
-router.delete('/me', async (req, res) => {
-  try {
-    await req.user.remove();
-    sendCancellationEmail(req.user.email, req.user.name);
-    res.clearCookie('jwt');
-    res.json({ message: 'user deleted' });
-  } catch (e) {
-    res.status(500).json({ error: e.toString() });
-  }
-});
+router.delete('/me', deleteUser);
 
 // ***********************************************//
 // Upload avatar
 // ***********************************************//
-router.post('/avatar', async (req, res) => {
-  try {
-    const response = await cloudinary.uploader.upload(
-      req.files.avatar.tempFilePath
-    );
-    req.user.avatar = response.secure_url;
-    await req.user.save();
-    res.json(response);
-  } catch (error) {
-    res.json({ error: e.toString() });
-  }
-});
+router.post('/avatar', uploadAvatar);
 
 // ******************************
 // Update password
 // ******************************
-router.put('/password', async (req, res) => {
-  try {
-    req.user.password = req.body.password;
-    await req.user.save();
-    res.clearCookie('jwt');
-    res.json({ message: 'password updated successfully' });
-  } catch (e) {
-    res.json({ error: e.toString() });
-  }
-});
+router.put('/password', updatePassword);
 
 module.exports = router;
